+++ src/main/java/com/ofds/dto/CartDTO.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 @Data
 public class CartDTO {
-    private Integer cartId;
-    private Integer id;
-    private Integer restaurantId;
+    private Long cartId; // Change the return type back to Integer if any issue occur.
+    private Long id;
+    private Long restaurantId;
     private String restaurantName;
     private Integer itemCount;
     private Double totalAmount;
