+++ src/main/java/com/ofds/dto/CartItemDTO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class CartItemDTO {
-    private Integer cartItemId;
-    private Integer menuItemId;
+    private Long cartItemId;
+    private Long menuItemId;
     private String name;
     private Double price;
     private Integer quantity;
