+++ src/main/java/com/ofds/repository/CartItemRepository.java
@@ -3,5 +3,5 @@ package com.ofds.repository;
 import com.ofds.entity.CartItemEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartItemRepository extends JpaRepository<CartItemEntity, Integer> {
+public interface CartItemRepository extends JpaRepository<CartItemEntity, Long> {
 }
